chore(contracts): clean up clippy warnings across workspace
Replaces deprecated Env::register_contract(None, X) calls with the
non-deprecated Env::register(X, ()) equivalent, applies clippy's
machine-applicable fixes (redundant casts, range-contains idioms,
unneeded match patterns, unused imports), and marks fraud-prevention's
mock cross-contract module #[allow(dead_code)] since its endpoints are
only invoked via generated Client calls, which rustc's dead-code check
can't see.

Leaves the remaining too-many-arguments warnings (22 call sites across
10 crates) untouched since fixing those needs real signature refactors,
not mechanical changes.

+++ contracts/analytics-aggregator/src/lib.rs
@@ -151,7 +151,7 @@ impl AnalyticsAggregatorContract {
                 / analytics.total_impressions as u128) as u32)
                 .min(10_000);
             // Use checked/saturating arithmetic for CPM
-            analytics.cpm = (analytics.total_spend as i128)
+            analytics.cpm = analytics.total_spend
                 .saturating_mul(1_000)
                 .checked_div(analytics.total_impressions as i128)
                 .unwrap_or(0);
